allow users to set ua arguments or a full ua object
Instead of exposing each property of LWP::UserAgent, it's
probably best to expose the entire building arguments used to
create the object. This allows us to set timeouts, env_proxy,
SSL/TLS options and more.

Another interesting and even more flexible approach is to allow
the user to pass on their own user agent object, as long as it's
HTTP::Request compatible.