format and unify changelog
The Changes file was reversed to the most recent changes appear first,
which conforms to the CPAN::Changes standard.

We also centralized all changelog into the Changes file, so we don't
have to update both there and on the module's POD.

Finally, we've also fixed a regression regarding the copyright year,
which started on 2012 and not on 2016.